% Generated in roxygen2: do not edit by hand
% Please edit documentation in R/avg_annual_rainfall.R

\name{avgrainfall}
\alias{avgrainfall}
\title{Average Annual Rainfall}
\usage{
avgrainfall()
}
\description{
Calculates the mean annual rainfall in US.
}
\examples{
avgrainfall()
}
\author{
Derek Nguyen and Jonathan Hart
}
